requirements.txt
youtube-transcript-api==0.6.2
chromadb==0.5.3
sentence-transformers==3.0.1
anthropic==0.34.0
python-dotenv==1.0.1
rich==13.7.1